Fixed Issues Regarding the Navbar
After taking a break from the code, I quickly found that I forgot to change the string on line 43 of App.js from Navbar to <Navbar /> (switching from string to component). I also found that one line was wrong on Navbar.jsx. On line 84, I had {isClicked.chatt && <Chat />} which blocked the application from knowing when the chat icon was clicked. I fixed it by making it {isClicked.chat && <Chat />}. Now the code seems to be working correctly up to this point with no errors that I can tell.