Added a Few Things to The Homepage
The homepage now contains an SVG image, as well as an earnings section with a download button. It also includes 4 cards, customers, products, sales, and refunds. Each card displays the title for the card with a corresponding image, it includes the data numbers for the specific card and includes a colored percentage gain or loss for what the card represents. If it is a net gain, the percentage is green, if it is a net loss, the percentage is red.

Everything seems to be working as it should with no errors at this point of the project.