Added the revenue section to the homepage
The revenue section of the homepage is now working as intended, as far as the project goes. I will need to come back and edit the data and the graphs to better represent the data. The project has it set up in a way that does not make sense, it looks good, but the stacked bar chart does not make sense as far as the course instructor instructed. I will need to come back eventually and make changes to them, but they work. No errors as far as I can tell.